requirements.txt
pytest
pytest-cov
sphinx
pyenchant
bs4
pyyaml
